+++ hledger-lib/Hledger/Data/Journal.hs
@@ -1,7 +1,5 @@
-{-# LANGUAGE LambdaCase          #-}
 {-# LANGUAGE NamedFieldPuns      #-}
 {-# LANGUAGE OverloadedStrings   #-}
-{-# LANGUAGE PackageImports      #-}
 {-# LANGUAGE Rank2Types          #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 
@@ -21,7 +19,6 @@ module Hledger.Data.Journal (
   addTransactionModifier,
   addPeriodicTransaction,
   addTransaction,
-  journalBalanceTransactions,
   journalInferMarketPricesFromTransactions,
   journalApplyCommodityStyles,
   commodityStylesFromAmounts,
@@ -86,34 +83,28 @@ module Hledger.Data.Journal (
   -- * Misc
   canonicalStyleFrom,
   nulljournal,
-  journalCheckBalanceAssertions,
+  journalNumberTransactions,
   journalNumberAndTieTransactions,
   journalUntieTransactions,
   journalModifyTransactions,
   journalApplyAliases,
   -- * Tests
   samplejournal,
-  tests_Journal,
+  samplejournalMaybeExplicit,
+  tests_Journal
 )
 where
 
 import Control.Applicative ((<|>))
-import Control.Monad.Except (ExceptT(..), runExceptT, throwError)
-import "extra" Control.Monad.Extra (whenM)
-import Control.Monad.Reader as R
-import Control.Monad.ST (ST, runST)
+import Control.Monad.Except (ExceptT(..))
 import Control.Monad.State.Strict (StateT)
-import Data.Array.ST (STArray, getElems, newListArray, writeArray)
 import Data.Char (toUpper, isDigit)
 import Data.Default (Default(..))
 import Data.Foldable (toList)
-import Data.Function ((&))
-import qualified Data.HashTable.Class as H (toList)
-import qualified Data.HashTable.ST.Cuckoo as H
-import Data.List ((\\), find, foldl', sortBy, sortOn)
+import Data.List ((\\), find, foldl', sortBy)
 import Data.List.Extra (nubSort)
 import qualified Data.Map.Strict as M
-import Data.Maybe (catMaybes, fromJust, fromMaybe, isJust, mapMaybe, maybeToList)
+import Data.Maybe (catMaybes, fromMaybe, mapMaybe, maybeToList)
 import qualified Data.Set as S
 import Data.Text (Text)
 import qualified Data.Text as T
@@ -129,7 +120,6 @@ import Hledger.Utils
 import Hledger.Data.Types
 import Hledger.Data.AccountName
 import Hledger.Data.Amount
-import Hledger.Data.Dates
 import Hledger.Data.Transaction
 import Hledger.Data.TransactionModifier
 import Hledger.Data.Posting
@@ -734,339 +724,6 @@ journalModifyTransactions d j =
       Right ts -> Right j{jtxns=ts}
       Left err -> Left err
 
--- | Check any balance assertions in the journal and return an error message
--- if any of them fail (or if the transaction balancing they require fails).
-journalCheckBalanceAssertions :: Journal -> Maybe String
-journalCheckBalanceAssertions = either Just (const Nothing) . journalBalanceTransactions defbalancingopts
-
--- "Transaction balancing", including: inferring missing amounts,
--- applying balance assignments, checking transaction balancedness,
--- checking balance assertions, respecting posting dates. These things
--- are all interdependent.
--- WARN tricky algorithm and code ahead. 
---
--- Code overview as of 20190219, this could/should be simplified/documented more:
---  parseAndFinaliseJournal['] (Cli/Utils.hs), journalAddForecast (Common.hs), journalAddBudgetGoalTransactions (BudgetReport.hs), tests (BalanceReport.hs)
---   journalBalanceTransactions
---    runST
---     runExceptT
---      balanceTransaction (Transaction.hs)
---       balanceTransactionHelper
---      runReaderT
---       balanceTransactionAndCheckAssertionsB
---        addAmountAndCheckAssertionB
---        addOrAssignAmountAndCheckAssertionB
---        balanceTransactionHelper (Transaction.hs)
---  uiCheckBalanceAssertions d ui@UIState{aopts=UIOpts{cliopts_=copts}, ajournal=j} (ErrorScreen.hs)
---   journalCheckBalanceAssertions
---    journalBalanceTransactions
---  transactionWizard, postingsBalanced (Add.hs), tests (Transaction.hs)
---   balanceTransaction (Transaction.hs)  XXX hledger add won't allow balance assignments + missing amount ?
-
--- | Monad used for statefully balancing/amount-inferring/assertion-checking
--- a sequence of transactions.
--- Perhaps can be simplified, or would a different ordering of layers make sense ?
--- If you see a way, let us know.
-type Balancing s = ReaderT (BalancingState s) (ExceptT String (ST s))
-
--- | The state used while balancing a sequence of transactions.
-data BalancingState s = BalancingState {
-   -- read only
-   bsStyles       :: Maybe (M.Map CommoditySymbol AmountStyle)  -- ^ commodity display styles
-  ,bsUnassignable :: S.Set AccountName                          -- ^ accounts in which balance assignments may not be used
-  ,bsAssrt        :: Bool                                       -- ^ whether to check balance assertions
-   -- mutable
-  ,bsBalances     :: H.HashTable s AccountName MixedAmount      -- ^ running account balances, initially empty
-  ,bsTransactions :: STArray s Integer Transaction              -- ^ a mutable array of the transactions being balanced
-    -- (for efficiency ? journalBalanceTransactions says: not strictly necessary but avoids a sort at the end I think)
-  }
-
--- | Access the current balancing state, and possibly modify the mutable bits,
--- lifting through the Except and Reader layers into the Balancing monad.
-withRunningBalance :: (BalancingState s -> ST s a) -> Balancing s a
-withRunningBalance f = ask >>= lift . lift . f
-
--- | Get this account's current exclusive running balance.
-getRunningBalanceB :: AccountName -> Balancing s MixedAmount
-getRunningBalanceB acc = withRunningBalance $ \BalancingState{bsBalances} -> do
-  fromMaybe nullmixedamt <$> H.lookup bsBalances acc
-
--- | Add this amount to this account's exclusive running balance.
--- Returns the new running balance.
-addToRunningBalanceB :: AccountName -> MixedAmount -> Balancing s MixedAmount
-addToRunningBalanceB acc amt = withRunningBalance $ \BalancingState{bsBalances} -> do
-  old <- fromMaybe nullmixedamt <$> H.lookup bsBalances acc
-  let new = maPlus old amt
-  H.insert bsBalances acc new
-  return new
-
--- | Set this account's exclusive running balance to this amount.
--- Returns the change in exclusive running balance.
-setRunningBalanceB :: AccountName -> MixedAmount -> Balancing s MixedAmount
-setRunningBalanceB acc amt = withRunningBalance $ \BalancingState{bsBalances} -> do
-  old <- fromMaybe nullmixedamt <$> H.lookup bsBalances acc
-  H.insert bsBalances acc amt
-  return $ maMinus amt old
-
--- | Set this account's exclusive running balance to whatever amount
--- makes its *inclusive* running balance (the sum of exclusive running
--- balances of this account and any subaccounts) be the given amount.
--- Returns the change in exclusive running balance.
-setInclusiveRunningBalanceB :: AccountName -> MixedAmount -> Balancing s MixedAmount
-setInclusiveRunningBalanceB acc newibal = withRunningBalance $ \BalancingState{bsBalances} -> do
-  oldebal  <- fromMaybe nullmixedamt <$> H.lookup bsBalances acc
-  allebals <- H.toList bsBalances
-  let subsibal =  -- sum of any subaccounts' running balances
-        maSum . map snd $ filter ((acc `isAccountNamePrefixOf`).fst) allebals
-  let newebal = maMinus newibal subsibal
-  H.insert bsBalances acc newebal
-  return $ maMinus newebal oldebal
-
--- | Update (overwrite) this transaction in the balancing state.
-updateTransactionB :: Transaction -> Balancing s ()
-updateTransactionB t = withRunningBalance $ \BalancingState{bsTransactions}  ->
-  void $ writeArray bsTransactions (tindex t) t
-
--- | Infer any missing amounts (to satisfy balance assignments and
--- to balance transactions) and check that all transactions balance
--- and (optional) all balance assertions pass. Or return an error message
--- (just the first error encountered).
---
--- Assumes journalInferCommodityStyles has been called, since those
--- affect transaction balancing.
---
--- This does multiple things at once because amount inferring, balance
--- assignments, balance assertions and posting dates are interdependent.
-journalBalanceTransactions :: BalancingOpts -> Journal -> Either String Journal
-journalBalanceTransactions bopts' j' =
-  let
-    -- ensure transactions are numbered, so we can store them by number
-    j@Journal{jtxns=ts} = journalNumberTransactions j'
-    -- display precisions used in balanced checking
-    styles = Just $ journalCommodityStyles j
-    bopts = bopts'{commodity_styles_=styles}
-    -- balance assignments will not be allowed on these
-    txnmodifieraccts = S.fromList . map (paccount . tmprPosting) . concatMap tmpostingrules $ jtxnmodifiers j
-  in
-    runST $ do
-      -- We'll update a mutable array of transactions as we balance them,
-      -- not strictly necessary but avoids a sort at the end I think.
-      balancedtxns <- newListArray (1, toInteger $ length ts) ts
-
-      -- Infer missing posting amounts, check transactions are balanced,
-      -- and check balance assertions. This is done in two passes:
-      runExceptT $ do
-
-        -- 1. Step through the transactions, balancing the ones which don't have balance assignments
-        -- and leaving the others for later. The balanced ones are split into their postings.
-        -- The postings and not-yet-balanced transactions remain in the same relative order.
-        psandts :: [Either Posting Transaction] <- fmap concat $ forM ts $ \case
-          t | null $ assignmentPostings t -> case balanceTransaction bopts t of
-              Left  e  -> throwError e
-              Right t' -> do
-                lift $ writeArray balancedtxns (tindex t') t'
-                return $ map Left $ tpostings t'
-          t -> return [Right t]
-
-        -- 2. Sort these items by date, preserving the order of same-day items,
-        -- and step through them while keeping running account balances,
-        runningbals <- lift $ H.newSized (length $ journalAccountNamesUsed j)
-        flip runReaderT (BalancingState styles txnmodifieraccts (not $ ignore_assertions_ bopts) runningbals balancedtxns) $ do
-          -- performing balance assignments in, and balancing, the remaining transactions,
-          -- and checking balance assertions as each posting is processed.
-          void $ mapM' balanceTransactionAndCheckAssertionsB $ sortOn (either postingDate tdate) psandts
-
-        ts' <- lift $ getElems balancedtxns
-        return j{jtxns=ts'}
-
--- | This function is called statefully on each of a date-ordered sequence of
--- 1. fully explicit postings from already-balanced transactions and
--- 2. not-yet-balanced transactions containing balance assignments.
--- It executes balance assignments and finishes balancing the transactions,
--- and checks balance assertions on each posting as it goes.
--- An error will be thrown if a transaction can't be balanced
--- or if an illegal balance assignment is found (cf checkIllegalBalanceAssignment).
--- Transaction prices are removed, which helps eg balance-assertions.test: 15. Mix different commodities and assignments.
--- This stores the balanced transactions in case 2 but not in case 1.
-balanceTransactionAndCheckAssertionsB :: Either Posting Transaction -> Balancing s ()
-balanceTransactionAndCheckAssertionsB (Left p@Posting{}) =
-  -- update the account's running balance and check the balance assertion if any
-  void . addAmountAndCheckAssertionB $ postingStripPrices p
-balanceTransactionAndCheckAssertionsB (Right t@Transaction{tpostings=ps}) = do
-  -- make sure we can handle the balance assignments
-  mapM_ checkIllegalBalanceAssignmentB ps
-  -- for each posting, infer its amount from the balance assignment if applicable,
-  -- update the account's running balance and check the balance assertion if any
-  ps' <- mapM (addOrAssignAmountAndCheckAssertionB . postingStripPrices) ps
-  -- infer any remaining missing amounts, and make sure the transaction is now fully balanced
-  styles <- R.reader bsStyles
-  case balanceTransactionHelper defbalancingopts{commodity_styles_=styles} t{tpostings=ps'} of
-    Left err -> throwError err
-    Right (t', inferredacctsandamts) -> do
-      -- for each amount just inferred, update the running balance
-      mapM_ (uncurry addToRunningBalanceB) inferredacctsandamts
-      -- and save the balanced transaction.
-      updateTransactionB t'
-
--- | If this posting has an explicit amount, add it to the account's running balance.
--- If it has a missing amount and a balance assignment, infer the amount from, and
--- reset the running balance to, the assigned balance.
--- If it has a missing amount and no balance assignment, leave it for later.
--- Then test the balance assertion if any.
-addOrAssignAmountAndCheckAssertionB :: Posting -> Balancing s Posting
-addOrAssignAmountAndCheckAssertionB p@Posting{paccount=acc, pamount=amt, pbalanceassertion=mba}
-  -- an explicit posting amount
-  | hasAmount p = do
-      newbal <- addToRunningBalanceB acc amt
-      whenM (R.reader bsAssrt) $ checkBalanceAssertionB p newbal
-      return p
-
-  -- no explicit posting amount, but there is a balance assignment
-  | Just BalanceAssertion{baamount,batotal,bainclusive} <- mba = do
-      newbal <- if batotal
-                   -- a total balance assignment (==, all commodities)
-                   then return $ mixedAmount baamount
-                   -- a partial balance assignment (=, one commodity)
-                   else do
-                     oldbalothercommodities <- filterMixedAmount ((acommodity baamount /=) . acommodity) <$> getRunningBalanceB acc
-                     return $ maAddAmount oldbalothercommodities baamount
-      diff <- (if bainclusive then setInclusiveRunningBalanceB else setRunningBalanceB) acc newbal
-      let p' = p{pamount=diff, poriginal=Just $ originalPosting p}
-      whenM (R.reader bsAssrt) $ checkBalanceAssertionB p' newbal
-      return p'
-
-  -- no explicit posting amount, no balance assignment
-  | otherwise = return p
-
--- | Add the posting's amount to its account's running balance, and
--- optionally check the posting's balance assertion if any.
--- The posting is expected to have an explicit amount (otherwise this does nothing).
--- Adding and checking balance assertions are tightly paired because we
--- need to see the balance as it stands after each individual posting.
-addAmountAndCheckAssertionB :: Posting -> Balancing s Posting
-addAmountAndCheckAssertionB p | hasAmount p = do
-  newbal <- addToRunningBalanceB (paccount p) $ pamount p
-  whenM (R.reader bsAssrt) $ checkBalanceAssertionB p newbal
-  return p
-addAmountAndCheckAssertionB p = return p
-
--- | Check a posting's balance assertion against the given actual balance, and
--- return an error if the assertion is not satisfied.
--- If the assertion is partial, unasserted commodities in the actual balance
--- are ignored; if it is total, they will cause the assertion to fail.
-checkBalanceAssertionB :: Posting -> MixedAmount -> Balancing s ()
-checkBalanceAssertionB p@Posting{pbalanceassertion=Just (BalanceAssertion{baamount,batotal})} actualbal =
-    forM_ (baamount : otheramts) $ \amt -> checkBalanceAssertionOneCommodityB p amt actualbal
-  where
-    assertedcomm = acommodity baamount
-    otheramts | batotal   = map (\a -> a{aquantity=0}) . amountsRaw
-                          $ filterMixedAmount ((/=assertedcomm).acommodity) actualbal
-              | otherwise = []
-checkBalanceAssertionB _ _ = return ()
-
--- | Does this (single commodity) expected balance match the amount of that
--- commodity in the given (multicommodity) actual balance ? If not, returns a
--- balance assertion failure message based on the provided posting.  To match,
--- the amounts must be exactly equal (display precision is ignored here).
--- If the assertion is inclusive, the expected amount is compared with the account's
--- subaccount-inclusive balance; otherwise, with the subaccount-exclusive balance.
-checkBalanceAssertionOneCommodityB :: Posting -> Amount -> MixedAmount -> Balancing s ()
-checkBalanceAssertionOneCommodityB p@Posting{paccount=assertedacct} assertedamt actualbal = do
-  let isinclusive = maybe False bainclusive $ pbalanceassertion p
-  actualbal' <-
-    if isinclusive
-    then
-      -- sum the running balances of this account and any of its subaccounts seen so far
-      withRunningBalance $ \BalancingState{bsBalances} ->
-        H.foldM
-          (\ibal (acc, amt) -> return $
-            if assertedacct==acc || assertedacct `isAccountNamePrefixOf` acc then maPlus ibal amt else ibal)
-          nullmixedamt
-          bsBalances
-    else return actualbal
-  let
-    assertedcomm    = acommodity assertedamt
-    actualbalincomm = headDef nullamt . amountsRaw . filterMixedAmountByCommodity assertedcomm $ actualbal'
-    pass =
-      aquantity
-        -- traceWith (("asserted:"++).showAmountDebug)
-        assertedamt ==
-      aquantity
-        -- traceWith (("actual:"++).showAmountDebug)
-        actualbalincomm
-
-    errmsg = printf (unlines
-                  [ "balance assertion: %s",
-                    "\nassertion details:",
-                    "date:       %s",
-                    "account:    %s%s",
-                    "commodity:  %s",
-                    -- "display precision:  %d",
-                    "calculated: %s", -- (at display precision: %s)",
-                    "asserted:   %s", -- (at display precision: %s)",
-                    "difference: %s"
-                  ])
-      (case ptransaction p of
-         Nothing -> "?" -- shouldn't happen
-         Just t ->  printf "%s\ntransaction:\n%s"
-                      (showGenericSourcePos pos)
-                      (textChomp $ showTransaction t)
-                      :: String
-                      where
-                        pos = baposition $ fromJust $ pbalanceassertion p
-      )
-      (showDate $ postingDate p)
-      (T.unpack $ paccount p) -- XXX pack
-      (if isinclusive then " (and subs)" else "" :: String)
-      assertedcomm
-      -- (asprecision $ astyle actualbalincommodity)  -- should be the standard display precision I think
-      (show $ aquantity actualbalincomm)
-      -- (showAmount actualbalincommodity)
-      (show $ aquantity assertedamt)
-      -- (showAmount assertedamt)
-      (show $ aquantity assertedamt - aquantity actualbalincomm)
-
-  unless pass $ throwError errmsg
-
--- | Throw an error if this posting is trying to do an illegal balance assignment.
-checkIllegalBalanceAssignmentB :: Posting -> Balancing s ()
-checkIllegalBalanceAssignmentB p = do
-  checkBalanceAssignmentPostingDateB p
-  checkBalanceAssignmentUnassignableAccountB p
-
--- XXX these should show position. annotateErrorWithTransaction t ?
-
--- | Throw an error if this posting is trying to do a balance assignment and
--- has a custom posting date (which makes amount inference too hard/impossible).
-checkBalanceAssignmentPostingDateB :: Posting -> Balancing s ()
-checkBalanceAssignmentPostingDateB p =
-  when (hasBalanceAssignment p && isJust (pdate p)) $
-    throwError . T.unpack $ T.unlines
-      ["postings which are balance assignments may not have a custom date."
-      ,"Please write the posting amount explicitly, or remove the posting date:"
-      ,""
-      ,maybe (T.unlines $ showPostingLines p) showTransaction $ ptransaction p
-      ]
-
--- | Throw an error if this posting is trying to do a balance assignment and
--- the account does not allow balance assignments (eg because it is referenced
--- by a transaction modifier, which might generate additional postings to it).
-checkBalanceAssignmentUnassignableAccountB :: Posting -> Balancing s ()
-checkBalanceAssignmentUnassignableAccountB p = do
-  unassignable <- R.asks bsUnassignable
-  when (hasBalanceAssignment p && paccount p `S.member` unassignable) $
-    throwError . T.unpack $ T.unlines
-      ["balance assignments cannot be used with accounts which are"
-      ,"posted to by transaction modifier rules (auto postings)."
-      ,"Please write the posting amount explicitly, or remove the rule."
-      ,""
-      ,"account: " <> paccount p
-      ,""
-      ,"transaction:"
-      ,""
-      ,maybe (T.unlines $ showPostingLines p) showTransaction $ ptransaction p
-      ]
-
 --
 
 -- | Choose and apply a consistent display style to the posting
@@ -1394,8 +1051,9 @@ journalApplyAliases aliases j =
 -- tracePostingsCommodities ps = trace (show $ map ((map (precision . acommodity) . amounts) . pamount) ps) ps
 
 -- tests
-
--- A sample journal for testing, similar to examples/sample.journal:
+--
+-- A sample journal for testing, similar to examples/sample.journal.
+-- Provide an option to either use explicit amounts or missing amounts, for testing purposes.
 --
 -- 2008/01/01 income
 --     assets:bank:checking  $1
@@ -1421,9 +1079,11 @@ journalApplyAliases aliases j =
 -- 2008/12/31 * pay off
 --     liabilities:debts  $1
 --     assets:bank:checking
---
-Right samplejournal = journalBalanceTransactions defbalancingopts $
-         nulljournal
+
+samplejournal = samplejournalMaybeExplicit True
+
+samplejournalMaybeExplicit :: Bool -> Journal
+samplejournalMaybeExplicit explicit = nulljournal
          {jtxns = [
            txnTieKnot $ Transaction {
              tindex=0,
@@ -1437,7 +1097,7 @@ Right samplejournal = journalBalanceTransactions defbalancingopts $
              ttags=[],
              tpostings=
                  ["assets:bank:checking" `post` usd 1
-                 ,"income:salary" `post` missingamt
+                 ,"income:salary" `post` if explicit then usd (-1) else missingamt
                  ],
              tprecedingcomment=""
            }
@@ -1454,7 +1114,7 @@ Right samplejournal = journalBalanceTransactions defbalancingopts $
              ttags=[],
              tpostings=
                  ["assets:bank:checking" `post` usd 1
-                 ,"income:gifts" `post` missingamt
+                 ,"income:gifts" `post` if explicit then usd (-1) else missingamt
                  ],
              tprecedingcomment=""
            }
@@ -1471,7 +1131,7 @@ Right samplejournal = journalBalanceTransactions defbalancingopts $
              ttags=[],
              tpostings=
                  ["assets:bank:saving" `post` usd 1
-                 ,"assets:bank:checking" `post` usd (-1)
+                 ,"assets:bank:checking" `post` if explicit then usd (-1) else missingamt
                  ],
              tprecedingcomment=""
            }
@@ -1488,7 +1148,7 @@ Right samplejournal = journalBalanceTransactions defbalancingopts $
              ttags=[],
              tpostings=["expenses:food" `post` usd 1
                        ,"expenses:supplies" `post` usd 1
-                       ,"assets:cash" `post` missingamt
+                       ,"assets:cash" `post` if explicit then usd (-2) else missingamt
                        ],
              tprecedingcomment=""
            }
@@ -1520,7 +1180,7 @@ Right samplejournal = journalBalanceTransactions defbalancingopts $
              tcomment="",
              ttags=[],
              tpostings=["liabilities:debts" `post` usd 1
-                       ,"assets:bank:checking" `post` usd (-1)
+                       ,"assets:bank:checking" `post` if explicit then usd (-1) else missingamt
                        ],
              tprecedingcomment=""
            }
@@ -1547,109 +1207,17 @@ tests_Journal = testGroup "Journal" [
       journalAccountNamesMatching :: Query -> Journal -> [AccountName]
       journalAccountNamesMatching q = filter (q `matchesAccount`) . journalAccountNames
       namesfrom qfunc = journalAccountNamesMatching (qfunc j) j
-    in [
-       testCase "assets"      $ assertEqual "" ["assets","assets:bank","assets:bank:checking","assets:bank:saving","assets:cash"]
+    in [testCase "assets"      $ assertEqual "" ["assets","assets:bank","assets:bank:checking","assets:bank:saving","assets:cash"]
          (namesfrom journalAssetAccountQuery)
-      ,testCase "cash"        $ assertEqual "" ["assets","assets:bank","assets:bank:checking","assets:bank:saving","assets:cash"]
-        (namesfrom journalCashAccountQuery)
-      ,testCase "liabilities" $ assertEqual "" ["liabilities","liabilities:debts"]
-        (namesfrom journalLiabilityAccountQuery)
-      ,testCase "equity"      $ assertEqual "" []
-        (namesfrom journalEquityAccountQuery)
-      ,testCase "income"      $ assertEqual "" ["income","income:gifts","income:salary"]
-        (namesfrom journalRevenueAccountQuery)
-      ,testCase "expenses"    $ assertEqual "" ["expenses","expenses:food","expenses:supplies"]
-        (namesfrom journalExpenseAccountQuery)
-    ]
-
-  ,testGroup "journalBalanceTransactions" [
-
-     testCase "balance-assignment" $ do
-      let ej = journalBalanceTransactions defbalancingopts $
-            --2019/01/01
-            --  (a)            = 1
-            nulljournal{ jtxns = [
-              transaction (fromGregorian 2019 01 01) [ vpost' "a" missingamt (balassert (num 1)) ]
-            ]}
-      assertRight ej
-      let Right j = ej
-      (jtxns j & head & tpostings & head & pamount & amountsRaw) @?= [num 1]
-
-    ,testCase "same-day-1" $ do
-      assertRight $ journalBalanceTransactions defbalancingopts $
-            --2019/01/01
-            --  (a)            = 1
-            --2019/01/01
-            --  (a)          1 = 2
-            nulljournal{ jtxns = [
-               transaction (fromGregorian 2019 01 01) [ vpost' "a" missingamt (balassert (num 1)) ]
-              ,transaction (fromGregorian 2019 01 01) [ vpost' "a" (num 1)    (balassert (num 2)) ]
-            ]}
-
-    ,testCase "same-day-2" $ do
-      assertRight $ journalBalanceTransactions defbalancingopts $
-            --2019/01/01
-            --    (a)                  2 = 2
-            --2019/01/01
-            --    b                    1
-            --    a
-            --2019/01/01
-            --    a                    0 = 1
-            nulljournal{ jtxns = [
-               transaction (fromGregorian 2019 01 01) [ vpost' "a" (num 2)    (balassert (num 2)) ]
-              ,transaction (fromGregorian 2019 01 01) [
-                 post' "b" (num 1)     Nothing
-                ,post' "a"  missingamt Nothing
-              ]
-              ,transaction (fromGregorian 2019 01 01) [ post' "a" (num 0)     (balassert (num 1)) ]
-            ]}
-
-    ,testCase "out-of-order" $ do
-      assertRight $ journalBalanceTransactions defbalancingopts $
-            --2019/1/2
-            --  (a)    1 = 2
-            --2019/1/1
-            --  (a)    1 = 1
-            nulljournal{ jtxns = [
-               transaction (fromGregorian 2019 01 02) [ vpost' "a" (num 1)    (balassert (num 2)) ]
-              ,transaction (fromGregorian 2019 01 01) [ vpost' "a" (num 1)    (balassert (num 1)) ]
-            ]}
-
-    ]
-
-    ,testGroup "commodityStylesFromAmounts" $ [
-
-      -- Journal similar to the one on #1091:
-      -- 2019/09/24
-      --     (a)            1,000.00
-      -- 
-      -- 2019/09/26
-      --     (a)             1000,000
-      --
-      testCase "1091a" $ do
-        commodityStylesFromAmounts [
-           nullamt{aquantity=1000, astyle=AmountStyle L False (Precision 3) (Just ',') Nothing}
-          ,nullamt{aquantity=1000, astyle=AmountStyle L False (Precision 2) (Just '.') (Just (DigitGroups ',' [3]))}
-          ]
-         @?=
-          -- The commodity style should have period as decimal mark
-          -- and comma as digit group mark.
-          Right (M.fromList [
-            ("", AmountStyle L False (Precision 3) (Just '.') (Just (DigitGroups ',' [3])))
-          ])
-        -- same journal, entries in reverse order
-      ,testCase "1091b" $ do
-        commodityStylesFromAmounts [
-           nullamt{aquantity=1000, astyle=AmountStyle L False (Precision 2) (Just '.') (Just (DigitGroups ',' [3]))}
-          ,nullamt{aquantity=1000, astyle=AmountStyle L False (Precision 3) (Just ',') Nothing}
-          ]
-         @?=
-          -- The commodity style should have period as decimal mark
-          -- and comma as digit group mark.
-          Right (M.fromList [
-            ("", AmountStyle L False (Precision 3) (Just '.') (Just (DigitGroups ',' [3])))
-          ])
-
-     ]
-
+       ,testCase "cash"        $ assertEqual "" ["assets","assets:bank","assets:bank:checking","assets:bank:saving","assets:cash"]
+         (namesfrom journalCashAccountQuery)
+       ,testCase "liabilities" $ assertEqual "" ["liabilities","liabilities:debts"]
+         (namesfrom journalLiabilityAccountQuery)
+       ,testCase "equity"      $ assertEqual "" []
+         (namesfrom journalEquityAccountQuery)
+       ,testCase "income"      $ assertEqual "" ["income","income:gifts","income:salary"]
+         (namesfrom journalRevenueAccountQuery)
+       ,testCase "expenses"    $ assertEqual "" ["expenses","expenses:food","expenses:supplies"]
+         (namesfrom journalExpenseAccountQuery)
+       ]
   ]
