+++ hledger-lib/Hledger/Data/Transaction.hs
@@ -7,69 +7,54 @@ tags.
 
 -}
 
-{-# LANGUAGE FlexibleContexts  #-}
 {-# LANGUAGE LambdaCase        #-}
 {-# LANGUAGE NamedFieldPuns    #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE Rank2Types        #-}
 {-# LANGUAGE RecordWildCards   #-}
-{-# LANGUAGE TemplateHaskell   #-}
 
-module Hledger.Data.Transaction (
-  -- * Transaction
-  nulltransaction,
-  transaction,
-  txnTieKnot,
-  txnUntieKnot,
-  transactionCheckBalanced,
+module Hledger.Data.Transaction
+( -- * Transaction
+  nulltransaction
+, transaction
+, txnTieKnot
+, txnUntieKnot
   -- * operations
-  showAccountName,
-  hasRealPostings,
-  realPostings,
-  assignmentPostings,
-  virtualPostings,
-  balancedVirtualPostings,
-  transactionsPostings,
-  BalancingOpts(..),
-  HasBalancingOpts(..),
-  defbalancingopts,
-  isTransactionBalanced,
-  balanceTransaction,
-  balanceTransactionHelper,
-  transactionTransformPostings,
-  transactionApplyValuation,
-  transactionToCost,
-  transactionApplyAliases,
-  transactionMapPostings,
-  transactionMapPostingAmounts,
-  -- nonzerobalanceerror,
+, showAccountName
+, hasRealPostings
+, realPostings
+, assignmentPostings
+, virtualPostings
+, balancedVirtualPostings
+, transactionsPostings
+, transactionTransformPostings
+, transactionApplyValuation
+, transactionToCost
+, transactionApplyAliases
+, transactionMapPostings
+, transactionMapPostingAmounts
+  -- nonzerobalanceerror
   -- * date operations
-  transactionDate2,
+, transactionDate2
   -- * transaction description parts
-  transactionPayee,
-  transactionNote,
-  -- payeeAndNoteFromDescription,
+, transactionPayee
+, transactionNote
+  -- payeeAndNoteFromDescription
   -- * rendering
-  showTransaction,
-  showTransactionOneLineAmounts,
-  -- showPostingLine,
-  showPostingLines,
+, showTransaction
+, showTransactionOneLineAmounts
+  -- showPostingLine
+, showPostingLines
   -- * GenericSourcePos
-  sourceFilePath,
-  sourceFirstLine,
-  showGenericSourcePos,
-  annotateErrorWithTransaction,
-  transactionFile,
+, sourceFilePath
+, sourceFirstLine
+, showGenericSourcePos
+, transactionFile
   -- * tests
-  tests_Transaction
-)
-where
+, tests_Transaction
+) where
 
 import Data.Default (Default(..))
-import Data.Foldable (asum)
-import Data.List (intercalate, partition)
-import Data.List.Extra (nubSort)
-import Data.Maybe (fromMaybe, isNothing, mapMaybe)
+import Data.Maybe (fromMaybe)
 import Data.Text (Text)
 import qualified Data.Text as T
 import qualified Data.Text.Lazy as TL
@@ -356,260 +341,6 @@ balancedVirtualPostings = filter isBalancedVirtual . tpostings
 transactionsPostings :: [Transaction] -> [Posting]
 transactionsPostings = concatMap tpostings
 
-data BalancingOpts = BalancingOpts
-  { ignore_assertions_ :: Bool  -- ^ Ignore balance assertions
-  , infer_transaction_prices_ :: Bool  -- ^ Infer prices in unbalanced multicommodity amounts
-  , commodity_styles_  :: Maybe (M.Map CommoditySymbol AmountStyle)  -- ^ commodity display styles
-  } deriving (Show)
-
-defbalancingopts :: BalancingOpts
-defbalancingopts = BalancingOpts
-  { ignore_assertions_ = False
-  , infer_transaction_prices_ = True
-  , commodity_styles_  = Nothing
-  }
-
--- | Check that this transaction would appear balanced to a human when displayed.
--- On success, returns the empty list, otherwise one or more error messages.
---
--- In more detail:
--- For the real postings, and separately for the balanced virtual postings:
---
--- 1. Convert amounts to cost where possible
---
--- 2. When there are two or more non-zero amounts
---    (appearing non-zero when displayed, using the given display styles if provided),
---    are they a mix of positives and negatives ?
---    This is checked separately to give a clearer error message.
---    (Best effort; could be confused by postings with multicommodity amounts.)
---
--- 3. Does the amounts' sum appear non-zero when displayed ?
---    (using the given display styles if provided)
---
-transactionCheckBalanced :: BalancingOpts -> Transaction -> [String]
-transactionCheckBalanced BalancingOpts{commodity_styles_} t = errs
-  where
-    (rps, bvps) = foldr partitionPosting ([], []) $ tpostings t
-      where
-        partitionPosting p ~(l, r) = case ptype p of
-            RegularPosting         -> (p:l, r)
-            BalancedVirtualPosting -> (l, p:r)
-            VirtualPosting         -> (l, r)
-
-    -- check for mixed signs, detecting nonzeros at display precision
-    canonicalise = maybe id canonicaliseMixedAmount commodity_styles_
-    signsOk ps =
-      case filter (not.mixedAmountLooksZero) $ map (canonicalise.mixedAmountCost.pamount) ps of
-        nonzeros | length nonzeros >= 2
-                   -> length (nubSort $ mapMaybe isNegativeMixedAmount nonzeros) > 1
-        _          -> True
-    (rsignsok, bvsignsok)       = (signsOk rps, signsOk bvps)
-
-    -- check for zero sum, at display precision
-    (rsum, bvsum)               = (sumPostings rps, sumPostings bvps)
-    (rsumcost, bvsumcost)       = (mixedAmountCost rsum, mixedAmountCost bvsum)
-    (rsumdisplay, bvsumdisplay) = (canonicalise rsumcost, canonicalise bvsumcost)
-    (rsumok, bvsumok)           = (mixedAmountLooksZero rsumdisplay, mixedAmountLooksZero bvsumdisplay)
-
-    -- generate error messages, showing amounts with their original precision
-    errs = filter (not.null) [rmsg, bvmsg]
-      where
-        rmsg
-          | rsumok        = ""
-          | not rsignsok  = "real postings all have the same sign"
-          | otherwise     = "real postings' sum should be 0 but is: " ++ showMixedAmount rsumcost
-        bvmsg
-          | bvsumok       = ""
-          | not bvsignsok = "balanced virtual postings all have the same sign"
-          | otherwise     = "balanced virtual postings' sum should be 0 but is: " ++ showMixedAmount bvsumcost
-
--- | Legacy form of transactionCheckBalanced.
-isTransactionBalanced :: BalancingOpts -> Transaction -> Bool
-isTransactionBalanced bopts = null . transactionCheckBalanced bopts
-
--- | Balance this transaction, ensuring that its postings
--- (and its balanced virtual postings) sum to 0,
--- by inferring a missing amount or conversion price(s) if needed.
--- Or if balancing is not possible, because the amounts don't sum to 0 or
--- because there's more than one missing amount, return an error message.
---
--- Transactions with balance assignments can have more than one
--- missing amount; to balance those you should use the more powerful
--- journalBalanceTransactions.
---
--- The "sum to 0" test is done using commodity display precisions,
--- if provided, so that the result agrees with the numbers users can see.
---
-balanceTransaction ::
-     BalancingOpts
-  -> Transaction
-  -> Either String Transaction
-balanceTransaction bopts = fmap fst . balanceTransactionHelper bopts
-
--- | Helper used by balanceTransaction and balanceTransactionWithBalanceAssignmentAndCheckAssertionsB;
--- use one of those instead. It also returns a list of accounts
--- and amounts that were inferred.
-balanceTransactionHelper ::
-     BalancingOpts
-  -> Transaction
-  -> Either String (Transaction, [(AccountName, MixedAmount)])
-balanceTransactionHelper bopts t = do
-  (t', inferredamtsandaccts) <- inferBalancingAmount (fromMaybe M.empty $ commodity_styles_ bopts) $
-    if infer_transaction_prices_ bopts then inferBalancingPrices t else t
-  case transactionCheckBalanced bopts t' of
-    []   -> Right (txnTieKnot t', inferredamtsandaccts)
-    errs -> Left $ transactionBalanceError t' errs
-
--- | Generate a transaction balancing error message, given the transaction
--- and one or more suberror messages.
-transactionBalanceError :: Transaction -> [String] -> String
-transactionBalanceError t errs =
-  annotateErrorWithTransaction t $
-  intercalate "\n" $ "could not balance this transaction:" : errs
-
-annotateErrorWithTransaction :: Transaction -> String -> String
-annotateErrorWithTransaction t s =
-  unlines [ showGenericSourcePos $ tsourcepos t, s
-          , T.unpack . T.stripEnd $ showTransaction t
-          ]
-
--- | Infer up to one missing amount for this transactions's real postings, and
--- likewise for its balanced virtual postings, if needed; or return an error
--- message if we can't. Returns the updated transaction and any inferred posting amounts,
--- with the corresponding accounts, in order).
---
--- We can infer a missing amount when there are multiple postings and exactly
--- one of them is amountless. If the amounts had price(s) the inferred amount
--- have the same price(s), and will be converted to the price commodity.
-inferBalancingAmount ::
-     M.Map CommoditySymbol AmountStyle -- ^ commodity display styles
-  -> Transaction
-  -> Either String (Transaction, [(AccountName, MixedAmount)])
-inferBalancingAmount styles t@Transaction{tpostings=ps}
-  | length amountlessrealps > 1
-      = Left $ transactionBalanceError t
-        ["can't have more than one real posting with no amount"
-        ,"(remember to put two or more spaces between account and amount)"]
-  | length amountlessbvps > 1
-      = Left $ transactionBalanceError t
-        ["can't have more than one balanced virtual posting with no amount"
-        ,"(remember to put two or more spaces between account and amount)"]
-  | otherwise
-      = let psandinferredamts = map inferamount ps
-            inferredacctsandamts = [(paccount p, amt) | (p, Just amt) <- psandinferredamts]
-        in Right (t{tpostings=map fst psandinferredamts}, inferredacctsandamts)
-  where
-    (amountfulrealps, amountlessrealps) = partition hasAmount (realPostings t)
-    realsum = sumPostings amountfulrealps
-    (amountfulbvps, amountlessbvps) = partition hasAmount (balancedVirtualPostings t)
-    bvsum = sumPostings amountfulbvps
-
-    inferamount :: Posting -> (Posting, Maybe MixedAmount)
-    inferamount p =
-      let
-        minferredamt = case ptype p of
-          RegularPosting         | not (hasAmount p) -> Just realsum
-          BalancedVirtualPosting | not (hasAmount p) -> Just bvsum
-          _                                          -> Nothing
-      in
-        case minferredamt of
-          Nothing -> (p, Nothing)
-          Just a  -> (p{pamount=a', poriginal=Just $ originalPosting p}, Just a')
-            where
-              -- Inferred amounts are converted to cost.
-              -- Also ensure the new amount has the standard style for its commodity
-              -- (since the main amount styling pass happened before this balancing pass);
-              a' = styleMixedAmount styles . mixedAmountCost $ maNegate a
-
--- | Infer prices for this transaction's posting amounts, if needed to make
--- the postings balance, and if possible. This is done once for the real
--- postings and again (separately) for the balanced virtual postings. When
--- it's not possible, the transaction is left unchanged.
---
--- The simplest example is a transaction with two postings, each in a
--- different commodity, with no prices specified. In this case we'll add a
--- price to the first posting such that it can be converted to the commodity
--- of the second posting (with -B), and such that the postings balance.
---
--- In general, we can infer a conversion price when the sum of posting amounts
--- contains exactly two different commodities and no explicit prices.  Also
--- all postings are expected to contain an explicit amount (no missing
--- amounts) in a single commodity. Otherwise no price inferring is attempted.
---
--- The transaction itself could contain more than two commodities, and/or
--- prices, if they cancel out; what matters is that the sum of posting amounts
--- contains exactly two commodities and zero prices.
---
--- There can also be more than two postings in either of the commodities.
---
--- We want to avoid excessive display of digits when the calculated price is
--- an irrational number, while hopefully also ensuring the displayed numbers
--- make sense if the user does a manual calculation. This is (mostly) achieved
--- in two ways:
---
--- - when there is only one posting in the "from" commodity, a total price
---   (@@) is used, and all available decimal digits are shown
---
--- - otherwise, a suitable averaged unit price (@) is applied to the relevant
---   postings, with display precision equal to the summed display precisions
---   of the two commodities being converted between, or 2, whichever is larger.
---
--- (We don't always calculate a good-looking display precision for unit prices
--- when the commodity display precisions are low, eg when a journal doesn't
--- use any decimal places. The minimum of 2 helps make the prices shown by the
--- print command a bit less surprising in this case. Could do better.)
---
-inferBalancingPrices :: Transaction -> Transaction
-inferBalancingPrices t@Transaction{tpostings=ps} = t{tpostings=ps'}
-  where
-    ps' = map (priceInferrerFor t BalancedVirtualPosting . priceInferrerFor t RegularPosting) ps
-
--- | Generate a posting update function which assigns a suitable balancing
--- price to the posting, if and as appropriate for the given transaction and
--- posting type (real or balanced virtual). If we cannot or should not infer
--- prices, just act as the identity on postings.
-priceInferrerFor :: Transaction -> PostingType -> (Posting -> Posting)
-priceInferrerFor t pt = maybe id inferprice inferFromAndTo
-  where
-    postings     = filter ((==pt).ptype) $ tpostings t
-    pcommodities = map acommodity $ concatMap (amounts . pamount) postings
-    sumamounts   = amounts $ sumPostings postings  -- amounts normalises to one amount per commodity & price
-
-    -- We can infer prices if there are no prices given, exactly two commodities in the normalised
-    -- sum of postings in this transaction, and these two have opposite signs. The amount we are
-    -- converting from is the first commodity to appear in the ordered list of postings, and the
-    -- commodity we are converting to is the other. If we cannot infer prices, return Nothing.
-    inferFromAndTo = case sumamounts of
-      [a,b] | noprices, oppositesigns -> asum $ map orderIfMatches pcommodities
-        where
-          noprices      = all (isNothing . aprice) sumamounts
-          oppositesigns = signum (aquantity a) /= signum (aquantity b)
-          orderIfMatches x | x == acommodity a = Just (a,b)
-                           | x == acommodity b = Just (b,a)
-                           | otherwise         = Nothing
-      _ -> Nothing
-
-    -- For each posting, if the posting type matches, there is only a single amount in the posting,
-    -- and the commodity of the amount matches the amount we're converting from,
-    -- then set its price based on the ratio between fromamount and toamount.
-    inferprice (fromamount, toamount) posting
-        | [a] <- amounts (pamount posting), ptype posting == pt, acommodity a == acommodity fromamount
-            = posting{ pamount   = mixedAmount a{aprice=Just conversionprice}
-                     , poriginal = Just $ originalPosting posting }
-        | otherwise = posting
-      where
-        -- If only one Amount in the posting list matches fromamount we can use TotalPrice.
-        -- Otherwise divide the conversion equally among the Amounts by using a unit price.
-        conversionprice = case filter (== acommodity fromamount) pcommodities of
-            [_] -> TotalPrice $ negate toamount
-            _   -> UnitPrice  $ negate unitprice `withPrecision` unitprecision
-
-        unitprice     = aquantity fromamount `divideAmount` toamount
-        unitprecision = case (asprecision $ astyle fromamount, asprecision $ astyle toamount) of
-            (Precision a, Precision b) -> Precision . max 2 $ saturatedAdd a b
-            _                          -> NaturalPrecision
-        saturatedAdd a b = if maxBound - a < b then maxBound else a + b
-
 -- Get a transaction's secondary date, defaulting to the primary date.
 transactionDate2 :: Transaction -> Day
 transactionDate2 t = fromMaybe (tdate t) $ tdate2 t
@@ -666,10 +397,6 @@ transactionFile Transaction{tsourcepos} =
   case tsourcepos of
     GenericSourcePos f _ _ -> f
     JournalSourcePos f _   -> f
-
--- lenses
-
-makeHledgerClassyLenses ''BalancingOpts
 
 -- tests
 
@@ -742,13 +469,6 @@ tests_Transaction =
             --       ["    a          $-0.01", "    b           $0.005", "    c           $0.005"]
 
             ]
-
-    , testCase "inferBalancingAmount" $ do
-         (fst <$> inferBalancingAmount M.empty nulltransaction) @?= Right nulltransaction
-         (fst <$> inferBalancingAmount M.empty nulltransaction{tpostings = ["a" `post` usd (-5), "b" `post` missingamt]}) @?=
-           Right nulltransaction{tpostings = ["a" `post` usd (-5), "b" `post` usd 5]}
-         (fst <$> inferBalancingAmount M.empty nulltransaction{tpostings = ["a" `post` usd (-5), "b" `post` (eur 3 @@ usd 4), "c" `post` missingamt]}) @?=
-           Right nulltransaction{tpostings = ["a" `post` usd (-5), "b" `post` (eur 3 @@ usd 4), "c" `post` usd 1]}
 
     , testGroup "showTransaction" [
           testCase "null transaction" $ showTransaction nulltransaction @?= "0000-01-01\n\n"
@@ -861,231 +581,5 @@ tests_Transaction =
                 , posting {paccount = "b", pamount = missingmixedamt}
                 ])) @?=
           (T.unlines ["2010-01-01 x", "    a          1 @ $2", "    b", ""])
-        ]
-    , testGroup "balanceTransaction" [
-         testCase "detect unbalanced entry, sign error" $
-          assertLeft
-            (balanceTransaction defbalancingopts
-               (Transaction
-                  0
-                  ""
-                  nullsourcepos
-                  (fromGregorian 2007 01 28)
-                  Nothing
-                  Unmarked
-                  ""
-                  "test"
-                  ""
-                  []
-                  [posting {paccount = "a", pamount = mixedAmount (usd 1)}, posting {paccount = "b", pamount = mixedAmount (usd 1)}]))
-        ,testCase "detect unbalanced entry, multiple missing amounts" $
-          assertLeft $
-             balanceTransaction defbalancingopts
-               (Transaction
-                  0
-                  ""
-                  nullsourcepos
-                  (fromGregorian 2007 01 28)
-                  Nothing
-                  Unmarked
-                  ""
-                  "test"
-                  ""
-                  []
-                  [ posting {paccount = "a", pamount = missingmixedamt}
-                  , posting {paccount = "b", pamount = missingmixedamt}
-                  ])
-        ,testCase "one missing amount is inferred" $
-          (pamount . last . tpostings <$>
-           balanceTransaction defbalancingopts
-             (Transaction
-                0
-                ""
-                nullsourcepos
-                (fromGregorian 2007 01 28)
-                Nothing
-                Unmarked
-                ""
-                ""
-                ""
-                []
-                [posting {paccount = "a", pamount = mixedAmount (usd 1)}, posting {paccount = "b", pamount = missingmixedamt}])) @?=
-          Right (mixedAmount $ usd (-1))
-        ,testCase "conversion price is inferred" $
-          (pamount . head . tpostings <$>
-           balanceTransaction defbalancingopts
-             (Transaction
-                0
-                ""
-                nullsourcepos
-                (fromGregorian 2007 01 28)
-                Nothing
-                Unmarked
-                ""
-                ""
-                ""
-                []
-                [ posting {paccount = "a", pamount = mixedAmount (usd 1.35)}
-                , posting {paccount = "b", pamount = mixedAmount (eur (-1))}
-                ])) @?=
-          Right (mixedAmount $ usd 1.35 @@ eur 1)
-        ,testCase "balanceTransaction balances based on cost if there are unit prices" $
-          assertRight $
-          balanceTransaction defbalancingopts
-            (Transaction
-               0
-               ""
-               nullsourcepos
-               (fromGregorian 2011 01 01)
-               Nothing
-               Unmarked
-               ""
-               ""
-               ""
-               []
-               [ posting {paccount = "a", pamount = mixedAmount $ usd 1 `at` eur 2}
-               , posting {paccount = "a", pamount = mixedAmount $ usd (-2) `at` eur 1}
-               ])
-        ,testCase "balanceTransaction balances based on cost if there are total prices" $
-          assertRight $
-          balanceTransaction defbalancingopts
-            (Transaction
-               0
-               ""
-               nullsourcepos
-               (fromGregorian 2011 01 01)
-               Nothing
-               Unmarked
-               ""
-               ""
-               ""
-               []
-               [ posting {paccount = "a", pamount = mixedAmount $ usd 1 @@ eur 1}
-               , posting {paccount = "a", pamount = mixedAmount $ usd (-2) @@ eur (-1)}
-               ])
-        ]
-    , testGroup "isTransactionBalanced" [
-         testCase "detect balanced" $
-          assertBool "" $
-          isTransactionBalanced defbalancingopts $
-          Transaction
-            0
-            ""
-            nullsourcepos
-            (fromGregorian 2009 01 01)
-            Nothing
-            Unmarked
-            ""
-            "a"
-            ""
-            []
-            [ posting {paccount = "b", pamount = mixedAmount (usd 1.00)}
-            , posting {paccount = "c", pamount = mixedAmount (usd (-1.00))}
-            ]
-        ,testCase "detect unbalanced" $
-          assertBool "" $
-          not $
-          isTransactionBalanced defbalancingopts $
-          Transaction
-            0
-            ""
-            nullsourcepos
-            (fromGregorian 2009 01 01)
-            Nothing
-            Unmarked
-            ""
-            "a"
-            ""
-            []
-            [ posting {paccount = "b", pamount = mixedAmount (usd 1.00)}
-            , posting {paccount = "c", pamount = mixedAmount (usd (-1.01))}
-            ]
-        ,testCase "detect unbalanced, one posting" $
-          assertBool "" $
-          not $
-          isTransactionBalanced defbalancingopts $
-          Transaction
-            0
-            ""
-            nullsourcepos
-            (fromGregorian 2009 01 01)
-            Nothing
-            Unmarked
-            ""
-            "a"
-            ""
-            []
-            [posting {paccount = "b", pamount = mixedAmount (usd 1.00)}]
-        ,testCase "one zero posting is considered balanced for now" $
-          assertBool "" $
-          isTransactionBalanced defbalancingopts $
-          Transaction
-            0
-            ""
-            nullsourcepos
-            (fromGregorian 2009 01 01)
-            Nothing
-            Unmarked
-            ""
-            "a"
-            ""
-            []
-            [posting {paccount = "b", pamount = mixedAmount (usd 0)}]
-        ,testCase "virtual postings don't need to balance" $
-          assertBool "" $
-          isTransactionBalanced defbalancingopts $
-          Transaction
-            0
-            ""
-            nullsourcepos
-            (fromGregorian 2009 01 01)
-            Nothing
-            Unmarked
-            ""
-            "a"
-            ""
-            []
-            [ posting {paccount = "b", pamount = mixedAmount (usd 1.00)}
-            , posting {paccount = "c", pamount = mixedAmount (usd (-1.00))}
-            , posting {paccount = "d", pamount = mixedAmount (usd 100), ptype = VirtualPosting}
-            ]
-        ,testCase "balanced virtual postings need to balance among themselves" $
-          assertBool "" $
-          not $
-          isTransactionBalanced defbalancingopts $
-          Transaction
-            0
-            ""
-            nullsourcepos
-            (fromGregorian 2009 01 01)
-            Nothing
-            Unmarked
-            ""
-            "a"
-            ""
-            []
-            [ posting {paccount = "b", pamount = mixedAmount (usd 1.00)}
-            , posting {paccount = "c", pamount = mixedAmount (usd (-1.00))}
-            , posting {paccount = "d", pamount = mixedAmount (usd 100), ptype = BalancedVirtualPosting}
-            ]
-        ,testCase "balanced virtual postings need to balance among themselves (2)" $
-          assertBool "" $
-          isTransactionBalanced defbalancingopts $
-          Transaction
-            0
-            ""
-            nullsourcepos
-            (fromGregorian 2009 01 01)
-            Nothing
-            Unmarked
-            ""
-            "a"
-            ""
-            []
-            [ posting {paccount = "b", pamount = mixedAmount (usd 1.00)}
-            , posting {paccount = "c", pamount = mixedAmount (usd (-1.00))}
-            , posting {paccount = "d", pamount = mixedAmount (usd 100), ptype = BalancedVirtualPosting}
-            , posting {paccount = "3", pamount = mixedAmount (usd (-100)), ptype = BalancedVirtualPosting}
-            ]
         ]
     ]
